+++ CHANGELOG.md
@@ -14,6 +14,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - removed mermaid2 support from the docs
 - added encoding types to all open commands
 - generate documentation images from code
+- added more input validations
+- added unit test for correct/least privilege permission
+- added check to prevent directory traversal
 
 ### Fixes